update process packing & invoice

+++ databases/postgre/models/sales.order.item.go
@@ -1,9 +1,9 @@
 package models
 
 type SalesOrderItem struct {
-	ID                 uint    `json:"id"`
-	SalesOrderId       uint    `json:"salesOrderId"`
-	ProductId          uint    `json:"productId"`
+	ID                 int64   `json:"id"`
+	SalesOrderId       int64   `json:"salesOrderId"`
+	ProductId          int64   `json:"productId"`
 	OrderQty           int     `json:"orderQty"`
 	BackOrderQty       int     `json:"backOrderQty"`
 	AllocationOrderQty int     `json:"allocationOrderQty"`
@@ -18,9 +18,9 @@ func (c *SalesOrderItem) TableName() string {
 }
 
 type SalesOrderItemRes struct {
-	ID                 uint    `json:"id"`
-	SalesOrderId       uint    `json:"salesOrderId"`
-	ProductId          uint    `json:"productId"`
+	ID                 int64   `json:"id"`
+	SalesOrderId       int64   `json:"salesOrderId"`
+	ProductId          int64   `json:"productId"`
 	Product            Product `json:"product" gorm:"foreignKey:ProductID"`
 	OrderQty           int     `json:"orderQty"`
 	BackOrderQty       int     `json:"backOrderQty"`
